done
added a many-to-many relation and new crud for the new entity

+++ rest-backend/src/main/java/com/example/restapi/Controller/OrderController.java
@@ -58,7 +58,7 @@ public class OrderController {
 
     @GetMapping("orders/filter/{price}")
     @ResponseStatus(HttpStatus.OK)
-    public List<EntityModel<Order>> filter_total(@PathVariable float price){
+    public List<EntityModel<Order>> filter_total(@PathVariable Double price){
         return this.service.filter_total(price);
     }
 }
